Remove 3000 simulation samples
This should have been done in the previous commit, where the idea was to replace 3000 samples (1000 per cell) that had a very slightly different estimate of the type-effect (in the 6th decimal or so). In the last commit I only added the correct simulation samples, but failed to remove the old ones. THis is what is done here.