Copy of simulations.R file
I create this copy so it can be used with the stored data files from the simulations ("simulations_stored.RData"). The simulations take a long time so I don't want to re-run them. However, they won't be uploaded to dataverse, so the dataverse script has to work on its on -- and we want to remove references to data files that don't exist, even if they are commented.